backend/requirements.txt
azure-functions==1.21.3
azure-functions-durable==1.5.0
azure-cosmos==4.9.0
azure-servicebus==7.12.3
azure-identity==1.19.0
azure-ai-projects==1.0.0b11
azure-search-documents==11.6.0
azure-keyvault-secrets==4.8.0
openai>=1.30.0
mcp==1.6.0
python-dotenv==1.0.1
